Renamed FooterBar to TabBar
